Split mmap_hashmap into multiple files
