+++ lib/common/common/src/lib.rs
@@ -26,9 +26,9 @@ pub mod math;
 pub mod maybe_uninit;
 pub mod memory_usage;
 pub mod mmap;
-pub mod mmap_hashmap;
 pub mod num_traits;
 pub mod panic;
+pub mod persisted_hashmap;
 pub mod process_counter;
 pub mod process_cpu_usage;
 pub mod progress_tracker;
